Split Deletion Modal to UserDeletionModal and ProjectDeletionModal
