+++ lib/lightning_web/live/profile_live/form_component.html.heex
@@ -1,7 +1,7 @@
 <div id={"user-#{@user.id}"} class="md:col-span-2">
   <%= if @action == :delete do %>
     <.live_component
-      module={LightningWeb.Components.DeletionModal}
+      module={LightningWeb.Components.UserDeletionModal}
       id={@user.id}
       user={@user}
       logout={true}
